use typescript in Preview.tsx
